Move entity dto to package id
