+++ src/main/java/com/gft/productapi/dto/id/ProductDto.java
@@ -1,9 +1,9 @@
-package com.gft.productapi.dto;
+package com.gft.productapi.dto.id;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter @Setter
-public class ClientDto {
+public class ProductDto {
     private Long id;
 }
