+++ src/main/java/com/gft/productapi/dto/request/SaleRequestDto.java
@@ -7,9 +7,9 @@ import java.util.Set;
 import javax.validation.constraints.NotNull;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.gft.productapi.dto.ClientDto;
-import com.gft.productapi.dto.SupplierDto;
-import com.gft.productapi.dto.ProductDto;
+import com.gft.productapi.dto.id.ClientDto;
+import com.gft.productapi.dto.id.SupplierDto;
+import com.gft.productapi.dto.id.ProductDto;
 
 import lombok.Getter;
 import lombok.Setter;
